Add several improvements
- Add new levels
- Add undo policy
- Divide level loading into 2 phases
- Develop settings layout
- Add level colors for showing difficulty
- Add icon
- Fix deffer requests

+++ app/src/main/java/spbpu/hsamcp/mathgame/level/Result.kt
@@ -1,6 +1,6 @@
 package spbpu.hsamcp.mathgame.level
 
-class Result(val steps: Int, val time: Long, val award: Award) {
+class Result(val steps: Float, val time: Long, val award: Award) {
     fun isBetter(other: Result?): Boolean {
         if (other == null) {
             return true
